narrowing warnings

+++ crypto_kem/hqc-128/clean/fft.c
@@ -17,7 +17,7 @@
 
 
 static void compute_fft_betas(uint16_t *betas);
-static void compute_subset_sums(uint16_t *subset_sums, const uint16_t *set, size_t set_size);
+static void compute_subset_sums(uint16_t *subset_sums, const uint16_t *set, uint16_t set_size);
 static void radix_t(uint16_t *f, const uint16_t *f0, const uint16_t *f1, uint32_t m_f);
 static void radix_t_big(uint16_t *f, const uint16_t *f0, const uint16_t *f1, uint32_t m_f);
 static void fft_t_rec(uint16_t *f, const uint16_t *w, size_t f_coeffs, uint8_t m, uint32_t m_f, const uint16_t *betas);
@@ -50,7 +50,7 @@ static void compute_fft_betas(uint16_t *betas) {
  * @param[in] set Array of set_size elements
  * @param[in] set_size Size of the array set
  */
-static void compute_subset_sums(uint16_t *subset_sums, const uint16_t *set, size_t set_size) {
+static void compute_subset_sums(uint16_t *subset_sums, const uint16_t *set, uint16_t set_size) {
     uint16_t i, j;
     subset_sums[0] = 0;
